Python 3.11
More patches to make installion more reliable

+++ RLBotGUIX main exe/RLBotGUIX.bat
@@ -2,20 +2,20 @@
 
 pushd "%localappdata%\RLBotGUIX"
 
-set rl_py="Python37\python.exe"
-set rl_pip="Python37\Scripts\pip.exe"
+set rl_py="Python311\python.exe"
+set rl_pip="Python311\Scripts\pip.exe"
 
-rem If Python 3.7.9 hasn't been installed yet, install it
+rem If Python 3.11.6 hasn't been installed yet, install it
 
 if not exist %rl_py% (
-  echo RLBot's Python 3.7.9 hasn't been installed! Please click either "Install" or "Repair", depending on what you see in the popup.
+  echo Installing/repairing RLBot's Python 3.11.6!
 
-  python-3.7.9-amd64.exe "TargetDir=%localappdata%\RLBotGUIX\Python37" Shortcuts=0 Include_launcher=0 InstallLauncherAllUsers=0 AssociateFiles=0 SimpleInstall=1 PrependPath=0
+  unzip.exe -qq python-3.11.6-custom-amd64.zip
 
-  echo Thank you for installing RLBot's Python 3.7.9.
+  echo Thank you for installing RLBot's Python 3.11.6
 )
 
-rem Since we have our own Python 3.7 install, we don't actually need to make a virtual environment
+rem Since we have our own Python 3.11 install, we don't actually need to make a virtual environment
 
 echo Checking for conenction to pypi.org...
 
@@ -24,12 +24,18 @@ rem We ping PyPi's package index to see if we have an internet connection, but d
 %WINDIR%\system32\ping -n 1 google.com > nul
 
 if %errorlevel% == 0 (
+  if not exist %rl_pip% (
+    echo Python pip not found - installing!
+    curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
+    %rl_py% get-pip.py
+  )
+
   echo Connected - Installing / upgrading RLBot components...
 
   %rl_py% -m pip install -U pip --no-warn-script-location
-  %rl_pip% install setuptools wheel --no-warn-script-location
-  %rl_pip% install gevent^<22 --no-warn-script-location
-  %rl_pip% install eel --no-warn-script-location
+  %rl_pip% install -U setuptools wheel --no-warn-script-location
+  %rl_pip% install -U gevent --no-warn-script-location
+  %rl_pip% install -U eel --no-warn-script-location
   %rl_pip% install -U rlbot_gui rlbot --no-warn-script-location
 ) else (
   echo.
